Delete cluster resources on stale node cluster
At the end of init, the reflectors will report which NodeClusters
currently exist on the system.  We use that as a trigger to delete stale
NodeCluster objects, but when we do so, we don't first try to delete any
related ClusterResource objects.  That leads to a DB constraint
violation because of the FK relationship.

Signed-off-by: Allain Legacy <alegacy@redhat.com>