Fix selector number handling (#542)
The response from the http handler needs to be marshaled to a map of
interface{} values so that it can be handled by the selector code.  The
JSON decoder treats any numeric value as a float64, therefore, we need
to change the selector evaluator to handle float64.

This wasn't needed in the original PoC implementation because the actual
object type was passed to the selector rather than a map of interface{}
values.  We cannot do that in this case because we only have the HTTP
ResponseWriter with the marshaled output rather than the actual object
type.

The behavior of the JSON marshaler is described here:

https://github.com/square/go-jose/pull/352

Signed-off-by: Allain Legacy <alegacy@redhat.com>