Update trust bundle procedure and naming (#553)
To better align with how trust bundles are created and used in other
parts of the system, the name of the expected file is being changed
from ca-bundle.pem to ca-bundle.crt.  OpenShift is inconsistent in how
it expects this file to be named in other areas. But, in the injected
trust bundle case, and in the cluster proxy case, the file is expected
to be ca-bundle.crt so we are aligning to that.

Beware that when providing a set of trusted client CA certificates to
the ingress controller, it is expected to be ca-bundle.pem.

Signed-off-by: Allain Legacy <alegacy@redhat.com>