Add OAuth security scheme to openapi specs
This adds a Security Scheme to our OpenAPI specification, which will
serve as documentation to indicate which scopes should be requested of
the authorization server to gain access to the API endpoint.  These
are specified as scopes in the token request.  The authorization server
admin will determine which scopes each client can request.  Each of
these scopes will translate to a specific role being assigned to the
token.

We are trying to promote a partial delegation model.  That is, the
authorization server admin determines which scopes a client may request,
while the Resource Server (i.e., our application) determines which API
endpoints each of those scopes will permit.  We do this by mapping
each scope to a Kubernetes role which we have previously defined [1].

The scope naming convention is the prefix "role:" followed by the
role name.  When this scope is defined in the authorization server it
will map to an action that adds the corresponding role to the token's
role list.

[1] config/rbac/oran_o2ims_user_roles.yaml

Signed-off-by: Allain Legacy <alegacy@redhat.com>