Replace kube-rbac-proxy with middleware
This adds middleware directly into the servers to handle authentication
and authorization.  This activity was mostly triggered by the
deprecation of the gcr.io kube-rbac-proxy image and the recommendation from
the kubebuilder teams to remove the use of the proxy in favor of
middleware.  Unfortunately, the middleware recommended was geared
directly towards metrics and wasn't generic enough to use as is, but this
code follows a similar approach with some customizations for our own
circumstances as well as additional code to support the OAuth/OIDC use
case.

Signed-off-by: Allain Legacy <alegacy@redhat.com>