CNF-16878: Add client certificate binding support
This adds support for client certificate bound access tokens (e.g.,
RFC8705).  When this feature is enabled on the authorization server
each token generated contains an attribute which represents the SHA256
fingerprint of the TLS certificate which was used by the client who
requested the token.

When this attribute is added to the token, Resource Servers are supposed
to validate that the token is only used by a client presenting that same
TLS certificate.  This is verified by comparing the fingerprint
contained within the token with the fingerprint of the certificate being
used to make the request.

We only support being behind an ingress proxy; therefore, we never
expect TLS to be terminated directly on each server for OAuth use cases.
For this reason, we only expect to get the client certificate information
from the proxy headers rather than from the local TLS stack.  Since
we only expect to be behind an HAProxy, we currently only support the
client certificate headers added by HAProxy and the proposed standard
of RFC9440.

Signed-off-by: Allain Legacy <alegacy@redhat.com>