Update TLS configuration parameters (#591)
Updating the Inventory CR definition so that it is possible to set a
custom TLS certificate for the ingress if necessary.  Since using
cert-manager + ACME to generate the TLS certificate to be used for
client TLS requires it to have our real ingress DNS domain, then we
should allow using it as the ingress TLS certificate.  This will
ensure that the SMO only needs to trust one certificate for both the
client and server role rather than potentially 2 different ones.  It
also decouples us from having to using the wildcard domain certificate
assigned to the default ingress controller if the application is to be
run on a custom/separate domain.

Signed-off-by: Allain Legacy <alegacy@redhat.com>