CNF-16291: Cluster Provisioning with metal3
Enable cluster provisioning with the Metal3 plugin by utilizing the
new hostRef field in the Cluster Instance object. This field allows
Bare Metal Hosts (BMH) and other node-level installation resources
(such as InfraEnv and NMStateConfig) to be created within the BMH
namespace. Consequently, the pull secret must be available in the
BMH namespace.

The existing logic that copies the pull secret from the template
namespace to the cluster namespace will be deprecated and removed
in the Metal3 repository when the corresponding update becomes
available.

Additionally, a new field has been introduced in the node object,
which introduces a dependency on the oran-hwmgr-plugin, as it
leverages the new field to determine and return the BMH namespace.

Signed-off-by: Tao Liu <tali@redhat.com>