CNF-17204: Enable Cluster Provisioning with metal3
This pull request adds support for cluster provisioning with the Metal3
plugin by utilizing the newly introduced hostRef field in the Cluster
instance object. The hostRef field enables the creation of Bare Metal
Hosts (BMH) and other node-level installation resources—such as InfraEnv
and NMStateConfig—within the BMH namespace. It leverages the unstructured
site-config cluster instance to eliminate dependencies on upgrading
the siteconfig package.

As a result of this change, the pull secret must now be available in
the BMH namespace. The current logic that copies the pull secret
from the template namespace to the cluster namespace will be
deprecated.

Additionally, this update introduces a new field in the Node object.
The plugin will utilize this new field to determine and return the
BMH namespace. For backward compatibility, the IMS will check for
the existing node label before transitioning to the new
Spec.HwMgrNodeNs field.

Signed-off-by: Tao Liu <tali@redhat.com>