Wait for BMH to become Available before powering off the host (#767)
When a node is deleted, the deallocation hook powers off the host as part
of the deprovisioning process. However, if a BIOS firmware update job
is still in progress at the time of deletion, the job may get stuck,
preventing the BMH from transitioning to the Available state.

This update ensures the host is powered off only after deprovisioning is
complete and the BMH has reached the Available state, allowing any pending
jobs to finish successfully. This behavior applies only if the BMH has not
been provisioned.

To protect provisioned hosts, deletion of a ProvisionRequest in the
Configuration In Progress state (i.e., after the host has been provisioned)
will now be blocked by a validating webhook.

Signed-off-by: Tao Liu <tali@redhat.com>