metal3: Fix handling of multiple matching nodes
The metal3 plugin should only be fetching matching nodes when initially
allocating nodes to a request. There may be multiple BMHs in the same
namespace with the same resource selector labels, so subsequent calls to
fetch matching nodes may return a different list.

In addition, update the ProvisioningRequest status calculation to
consider the HardwareConfigured condition to handle failures.

Signed-off-by: Don Penney <dpenney@redhat.com>